Update 48x48 refresh icon
The previous version of this icon had a different resolution (pixels/mm)
than the other icons. This caused it to be displayed at a different size
when used in the main toolbar.

Additionally, the IconResourceLoader::LoadGdiplusBitmapFromPNGAndScale
function has been updated to call SetResolution(), which should help to
prevent issues like this in the future.