Add custom size gripper control
This control simply paints the size gripper shown in the bottom right
corner of resizeable dialogs.

This has been done because the built-in SCROLLBAR control (which is
what's normally used for the gripper) seemingly doesn't provide a way to
customise its background color. When using the standard theme, that's
fine. However, in dark mode, the background color of the scrollbar is
fixed to a lighter color than what's being used in the dialogs.

That color seems to be designed to match the color of the footer in the
common file (i.e. open/save) dialogs. Using that color as the background
color in the dialogs here tends to make them look washed out and not as
dark. That's why the regular SCROLLBAR control isn't being used.

Using a custom control works out ok in this instance because the size
gripper is essentially a window decoration. It doesn't change appearance
in any way and doesn't process any user input.

Note that the gripper window is now created programmatically. This has
the advantage that the position is always fixed to the bottom right
corner of the dialog and the control doesn't have to appear in each
dialogs resource.

There were also two other alternatives considered:

- Drawing the gripper onto the dialog background via WM_PAINT. Because
  the gripper is really just a decoration, it is possible to simply
  paint it on to the dialog background. However, that's overkill, as the
  entire background would have to be redrawn and invalidated properly.

- Custom drawing the SCROLLBAR control (via WM_PAINT). A SCROLLBAR
  control can have various styles, that determine whether it's drawn as
  a traditional scrollbar or as a gripper. When it is drawn as a
  gripper, the gripper can be styled in different ways (e.g. styled for
  the top left corner or top right corner).

  Handling those styles when painting would be unnecessary, but not
  handling them would mean the control couldn't be used as a generic
  scrollbar anymore.

  A custom control has the advantage that it's only designed to be used
  in one specific place and is explicitly not generic.

This is part of the work for issue #115.