Update toolbar in manage bookmarks dialog for dark mode
A static control is now set as the parent of the toolbar. For whatever
reason, the background color of the toolbar only appears to be set in
certain situations. For example, the main toolbars are contained within
a rebar and setting the background color of the rebar results in that
color being shown behind the toolbars.

In a similar way, the tab bar toolbar is contained within a static
control and setting the background color of that static control results
in the color being shown behind the toolbar.

In this case, the toolbar was a direct child of the dialog and although
the dialog had its background color set, the toolbar still had the
original background color. It could be that there's a particular window
message that needs to be handled in order to ensure that the toolbar
background color is correctly set.

In lieu of knowledge about that, it's easy enough to just place the
toolbar within a static control that has its background color set.

As part of this change, the text and button highlight colors are also
set for the toolbar, in the same way as they are for the main toolbars.

This is part of the work for issue #115.