Fix standards conformance issue
EMPTY_STRING is const, while the pszText member of the TBBUTTONINFO
struct isn't. Therefore, it's not valid for pszText to point at
EMPTY_STRING.

Note, however, that in practice, TB_SETBUTTONINFO will only read data
from the TBBUTTONINFO struct, so it would be ok to use const_cast. Only
setting the text when necessary is easier, however, so that's what's
done here.