Declare two window procedure stubs as static functions
Previously, these were declared as friend functions. However, the only
declarations for the functions were the friend declarations. That could
then cause lookup issues if attempting to compile in standards
conforming mode.

It's possible to fix those issues just by adding declarations for the
functions outside of the relevant classes, but as static window
procedure stubs are used elsewhere, that's what's used in these cases
now.