Invert color of Windows 10 icons in dark mode
This isn't perfect (the icons weren't designed to be shown inverted),
but when the icons are shown normally, they're not very visible on the
dark backgrounds that are used. So this works out much better overall.

This was suggested in https://github.com/derceg/explorerplusplus/issues/115#issuecomment-641959149.

This is part of the work for issue #115.