Tweak layout of general pane in options dialog
The language option has been moved into the general settings group.
Also, there is now a note identifier beside the language and icon theme
settings that indicates that, for the relevant changes to take effect,
the application needs to be restarted when either one of the settings is
changed.

Previously, there was some text to indicate this next to the language
setting, but there was nothing to indicate that the application also
needs to be restarted to change the icon theme.