Fix crash when closing selected tab
When the selected tab is closed and there is no previously selected tab
(i.e. because only a single tab has ever been selected), the selection
is transferred to the next tab. The only exception to this is when the
last tab is closed, in which case the selection is transferred to the
previous tab.

Previously, the selection wasn't been set properly in the usual case. If
the last tab was closed, the selection would be correctly set to the
previous tab. However, if any tab but the last was closed, the selection
would be kept as-is and the tab would be closed.

This would then result in a situation in which no tab was selected,
which would cause the application to crash, due to the assumption that
there's always a selected tab being broken.

To fix this, the next tab is now selected before a tab (not in the last
position) is closed.

This fixes #252.