Show error message when opening item on application toolbar fails
In previous versions of Explorer++, it wasn't necessary to quote a path
with spaces when adding a button to the application toolbar. It now is
necessary. That means that attempting to open an application toolbar
item created in a previous version of Explorer++ may fail (due to an
unquoted path).

To help make it clearer what's going wrong when an item fails to be
opened, the associated error message is now displayed. The parsed
application path is also shown. This should make it more clear that
quotes are required for paths with spaces.

See #145.

(cherry picked from commit 58e9a217e1e9f5d119eccb5c778832032bfa144e)