Fix crash when selecting address bar with specific toolbar configuration
When the focus is switched between the listview/treeview/address bar,
the main toolbar state is updated. Previously, the AddressBar class
accepted a pointer to the MainToolbar instance and called a method on
that instance directly when the focus changed.

However, because the toolbars are initialized in the order the appear,
if the main toolbar was moved below the address bar, the MainToolbar
instance would no longer be initialized before the AddressBar instance,
so the pointer would be invalid. Switching focus to the address bar
would then cause a crash.

To fix this, the main toolbar now updates its own state in response to a
focus change event, rather than having callers do the work.

This issue was reported at
https://explorerplusplus.com/forum/viewtopic.php?f=5&t=2727.

Additionally, focus will now be set back to the address bar if that was
the last window that had focus. Previously, focus would always go back
to the treeview/listview (whichever had focus before the address bar was
selected).