Destroy taskbar tab proxies on application shutdown
When the application is closed, DestroyWindow() is called on the main
window. Each child window is then destroyed before the main window is
destroyed.

When a particular child window is destroyed, any associated objects may
be destroyed as well. For example, when the tab control is destroyed,
the TabContainer instance is destroyed as well.

During shutdown, it's assumed that messages won't be interleaved with
WM_NCDESTROY. That is, once WM_NCDESTROY is sent to a child window and
that window is destroyed, it's implicitly assumed that the same process
continues up through the window chain.

The reason this is implicitly assumed is that various window message
handlers assume that key components (such as the TabContainer instance)
exist. Therefore, handling a message after destruction has started could
cause issues.

Because the taskbar tab proxies are top-level windows, they're not
directly affected by the destruction of the main window. Therefore,
normal window messages can still be sent to the tab proxies after
destruction of the main window has started.

This could cause a crash during application shutdown. The first thing
TaskbarThumbnails::TabProxyWndProc() does is retrieve the associated
tab. This won't work if the tab control and TabContainer instance have
already been destroyed.

To fix this, the tab proxies are now destroyed after the application is
guaranteed to close, but before the window destruction has started. This
means the tab proxies won't exist at the point at which the tab control
is destroyed, so they can't receive any window messages.