Fix double free of image list
By default, a listview will destroy its image list when the control is
destroyed. In the manage bookmarks dialog, that's undesirable. The
listview that's shown has an image list managed by BookmarkIconManager.
When the BookmarkIconManager instance is destroyed, it will destroy the
image list. Therefore, if the listview destroys the image list, the
BookmarkIconManager instance will attempt to do it a second time
afterwards.

To fix this, the LVS_SHAREIMAGELISTS style is now set on the listview.
This prevents the listview from destroying the image list.

This issue was found with Application Verifier.

(cherry picked from commit 6f04d291217613ed1f909dd41390df9d47424a5d)