Fix a large number of leaks in XML load/save code
A lot of items in the XML load/save functions were being leaked.
Previously, this didn't create too much of a practical issue, as data
was only loaded once and saved once.

However, application settings are now saved every 30 seconds. This means
that if any of the XML save functions, in particular, leak data, a
significant amount of memory can be leaked over time. Therefore, it's
critical that the functions release all memory, as necessary.

To fix this, WIL is now used to manage resources in the XML load/save
functions.

It's quite possible that this behaviour was the cause of
https://explorerplusplus.com/forum/viewtopic.php?f=2&t=2732 and
https://explorerplusplus.com/forum/viewtopic.php?f=2&t=2752.