Correct some usages of wil::com_ptr
wil::com_ptr has a constructor that accepts a raw pointer of the
appropriate type. This constructor then calls AddRef() on the COM
object. In other words, this constructor adds an additional reference to
the object.

Within the code here, this behaviour is undesirable. When a raw pointer
is wrapped in a wil::com_ptr, the intended behaviour is for the
wil::com_ptr to take ownership of the raw pointer (i.e. the wil::com_ptr
object shouldn't call AddRef()).

To accomplish that, the attach() method provided by the wil::com_ptr
class can be used. That method doesn't increment the reference count of
the COM object at all.

Therefore, in places where a wil::com_ptr object was being directly
constructed with a raw pointer, attach() is now called instead. This
should prevent the underlying COM objects from leaking.