Tweak allowable beta tag naming schemes
On AppVeyor, a beta build is detected by examining the name of the tag
(during a tag build). Previously, the tag would need to end in "-beta".
However, since a second beta for 1.4.0 is now planned, the naming has
been tweaked to allow tag names of the form "-beta(-\d)?".