Make dialog labels wider, where possible
When using a translation, the text within a dialog may be cut off (if
the translated text is longer than the original English text). The best
way of handling that is to manually resize the dialogs to fit the
translated text. However, doing that is tedious and doesn't really fit
with the way translations are managed within Crowdin.

When a translation is exported from Crowdin, the resource file that's
generated is the same as the English resource file, except that the
English strings are replaced with the translated strings.

That means that even if the dialogs in the translated resource file were
resized, those changes would be lost when regenerating the translated
files.

One way of attempting to ensure that there's sufficient space for the
translated strings is simply to make the labels within each dialog
wider and that's what's been done here.

Note that this is imperfect, since not all the controls in each dialog
have been resized (e.g. buttons have been left as-is). Additionally,
even if individual controls have been resized, text still might be cut
off because the parent dialog isn't large enough. There's no simple way
of handling that (apart from perhaps making the dialogs so large, they
will likely be large enough to contain any translated text).

Therefore, the changes here are meant to improve the situation, but not
completely fix it.

An example of this issue was reported in
https://explorerplusplus.com/forum/viewtopic.php?f=5&t=2768. Also see
issue #174.