Switch to editing name when renaming an item
An item can have several different types of names. For example, a name
used for display and a name used for parsing (see
https://docs.microsoft.com/en-us/windows/win32/api/shobjidl_core/ne-shobjidl_core-_shgdnf).

The different names don't necessarily have to be the same. For instance,
the C:\ drive item may have the following names:

- SHGDN_FORPARSING: C:\
- SHGDN_INFOLDER: Local Disk (C:)
- SHGDN_INFOLDER | SHGDN_FOREDITING: Local Disk

Therefore, when editing an item, it can be important to switch from the
display name (represented by SHGDN_INFOLDER) to the editing name
(represented by SHGDN_INFOLDER | SHGDN_FOREDITING). This is what's done
now.

A complication to this has to do with file extensions. Depending on
whether file extensions are hidden or shown in Windows Explorer, editing
names may or may not contain extensions. Additionally, extensions can be
independently hidden or shown in Explorer++.

Therefore, the editing name won't be used if the following conditions
are met:

- The item is a file
- Extensions are shown within Explorer++ and the only difference between
  the editing name and the display name is that the editing name is
  missing an extension, or
- Extensions are hidden within Explorer++ and the only difference
  between the editing name and the display name is that the editing name
  contains an extension.

In all other cases, the editing name will be used.