Update way in which items are renamed
Previously, when renaming an item, IFileOperation was being used. Whilst
that works for most items, it doesn't work for drives (where renaming
the drive ultimately just changes the drive label).

To fix that, IShellFolder::SetNameOf() is now used instead. That method
is less flexible than the corresponding IFileOperation functionality,
but allows drives to be renamed, in addition to other items (and the
additional flexibility of IFileOperation isn't really needed when
renaming anyway).

Along with this change, a few different checks and custom errors have
been removed from Explorerplusplus::OnListViewEndLabelEdit():

- There's no need to check whether a name ends in one or more '.'
  characters. The shell will simply strip them off anyway.
- There's no need to check for invalid characters.
  IShellFolder::SetNameOf will show a dedicated error message if an
  invalid character is used.
- There's no error message displayed by the application if renaming
  fails. The shell should display an error in that case.

One other temporary consequence of this change is that renames aren't
being recorded by the application for undo/redo. Also, the mass rename
dialog is still currently using the IFileOperation interface for
renames.