Rewrite function used to retrieve directory size using std::filesystem
The previous version of the function used
FindFirstFile()/FindNextFile(). Based on some user submitted crash
dumps, it appears the function would sometimes enter an infinite
recursive loop, eventually causing the application to crash.

The function used to assume that the first file returned by
FindFirstFile() would be ".". However, if the "." file was actually
returned by FindNextFile(), the function would recursively call itself
on the same directory over and over, and eventually run out of stack
space and crash. From the user submitted crash dumps, it appears this is
what was happening.

Because std::filesystem::directory_iterator skips over "." and "..",
that specific issue shouldn't be a problem anymore. Additionally, the
new function is simpler and is essentially a drop-in replacement.