Remove play navigation sound functionality
This functionality was similar to some functionality in Windows
Explorer, where a sound can be played when navigating into a folder.
However, the functionality here was more restricted. Although the sound
could be turned on or off, it couldn't be changed to an alternative
sound.

This functionality has now been removed. Sounds aren't played in any
other situations and playing a sound on navigation isn't very useful
(and multiple users have asked about turning it off before).

This effectively closes #268.