Remove call to FindFirstFile() during enumeration
Previously, FindFirstFile() was being used to retrieve basic information
on filesystem items. However, calling that method repeatedly (as opposed
to calling FindFirstFile() followed by repeated calls to FindNextFile())
causes excessively poor performance. This poor performance was more
noticeable on larger folders and networked folders.

To fix that issue, IShellFolder2::GetDetailsEx() is now used to retrieve
the same information that would be retrieved by FindFirstFile(). This
performs markedly better.

Over four runs whilst debugging, the approximate time taken to run
ShellBrowser::EnumerateFolder() on a local folder with 25000 items was
recorded. The average time under the old implementation and new
implementation is shown below:

Old: ~8.5s
New: ~1.9s

While the specific time taken to enumerate a folder will always vary,
it's consistently quicker to call IShellFolder2::GetDetailsEx() for each
item than to call FindFirstFile(). This likely has to do with the fact
IShellFolder2::GetDetailsEx() retrieves data cached within an item's
PIDL, which means the data only has to be retrieved once.

Although it's not necessarily an authoritative source, the Wine source
code does provide some indication of the data that's cached within the
PIDL structure:

https://github.com/wine-mirror/wine/blob/e909986e6ea5ecd49b2b847f321ad89b2ae4f6f1/dlls/shell32/pidl.h#L184

It's worth noting here that an alternative method that can be used to
query basic file information is SHGetDataFromIDList(). However that
method isn't used here, since it seemingly doesn't work for library
folders.

From debugging into SHGetDataFromIDList() on Windows 10, it appears that
it calls IShellFolder2::GetDetailsEx() with PKEY_FindData. It's this
call that then fails, which is something discussed at
https://social.msdn.microsoft.com/Forums/sharepoint/en-US/4d9f8a9e-0da8-4754-82da-cbdd86d9ccbb/ishellfolder2getdetailsex-with-pkeyfinddata-returns-einvalidarg-on-files-inside-document.

Using IShellFolder2::GetDetailsEx() directly is more complicated, but at
least consistent, since it can be used for every folder. Additionally,
whether PKEY_FindData is used or one of the individual property keys,
the information is still retrieved from data cached in the PIDL.

One thing worth pointing out is that IShellFolder2::GetDetailsEx() isn't
implemented in Wine, while SHGetDataFromIDList() is. If necessary,
SHGetDataFromIDList() could be used by default and
IShellFolder2::GetDetailsEx() could be used as a fallback.

One disadvantage is that the complexity of the change is enough that
it's hard to verify that there are no behavioural changes between the
old and new implementation. They should give the same results, but it's
possible there may be some subtle unintended differences.

This closes #262.