Fix issue with selecting current folder when navigating up
The code to do this previously assumed that once the initial navigation
call had completed, the navigation would be finished. As navigation is
now asynchronous, this is no longer the case. Therefore, the selection
will now be changed in a callback once navigation has completed.

This is part of the work for issue #283.