Explicitly set wait cursor when starting enumeration
During an enumeration, the wait cursor will be set by handling the
WM_SETCURSOR message. However, that message may only be sent when the
mouse moves. Therefore, the cursor is explicitly reset when the
enumeration is finished (in case the mouse doesn't move).

The cursor is now also explicitly set when the enumeration starts. This
means that cursor will change, even if the mouse doesn't move.

This is part of the work for issue #283.