Only set status bar text when menu isn't open
When a menu (like the main menu) is open, some helper text will be shown
in the status bar. Because folders are now loaded asynchronously, a load
can finish while a menu is being shown. If that's the case, the menu
helper text should still be shown in the status bar.

Additionally, when a menu is closed, the status bar text will now be
updated. This is done because if a folder is loaded while a menu is
being shown, the text in the status bar will be stale by the time the
menu is closed.

This is part of the work for issue #283.