Fix issue with showing properties for directory
ShellBrowser::ShowPropertiesForSelectedFiles() can be invoked from the
right-click background menu. When that happens, no files are selected
and the properties dialog is shown for the current folder.

In commit 7c3fcb7b65ef6e9c79c1c396f7cb8d3c3bc19709, the function was
updated to return early if the set of item PIDLs was empty. However, for
the reason given above, it's valid for the function to be invoked even
if no items are selected.

Therefore, the check has now been removed.