Only set/reset cursor when it's within listview
Previously, the cursor was being set and reset (when enumeration
started/finished), no matter where the cursor was located. Now, it's
only done if the cursor is within the listview (but not within the
listview header).

This matches the behaviour of the WM_SETCURSOR handler, where the cursor
is only set when the cursor is within the listview.

This is part of the work for issue #283.