Add option to check value of System.IsPinnedToNameSpaceTree property
The value of this property indicates whether a folder is pinned to the
navigation pane (in Windows Explorer). An option has now been added to
check this property when populating the treeview. To keep the original
behaviour, the option is disabled by default. If the user enables the
option and the property isn't set for a folder, that folder will be
skipped over when inserting items into the treeview.

Note that this option is available to be set via a new advanced pane
that has been added to the options dialog. This has been done for a
couple of reasons:

- The option is relatively obscure and very specific, so marking it as
  an advanced option makes more sense.
- Even if it weren't an advanced option, there's no space for the option
  in the only other place that would make sense (the "Treeview" section
  in the window pane).
- Other options can be added to the advanced pane. For example, the
  option to enable dark mode.

This closes #95.