Promote dark mode setting from command line parameter to advanced option
The dark mode functionality is incomplete, but it likely works well
enough that having a dedicated option makes sense. Therefore, the
command line option has been removed and a new advanced option has been
added. This also makes the option more visible.

Note that extra command line options are now ignored, so that the
presence of the previous option doesn't cause the application to
immediately exit.

One potential issue with this change is that if a future version of
Windows 10 changes the ordinal numbers assigned to the dark mode
functions or changes the dark mode functions themselves, the application
could crash on startup, at least until an update is made available.

A command line option is something the user has to explicitly set, so
it's easy for them to remove if it seems to cause the application to
crash. Changing a setting in the registry/config file might be a bit
more work, though.

The only potential solution to that might be to do strict version
checks, but then the dark mode functionality wouldn't work on newer
versions of Windows 10, again at least until an update is made
available.

This is part of the work for #115.