Update method used to retrieve basic file information
Previously, IShellFolder2::GetDetailsEx() was being used to retrieve
basic file information in ShellBrowser::SetItemInformation(). However,
that method was failing for some items in the desktop root folder.

Specifically, the function would fail to retrieve file information for
the user folder that appears on the desktop. It would also fail for the
OneDrive folder.

To fix that, a combination of SHGetDataFromIDList() and
IPropertyStore::GetValue() are now used to retrieve file properties. If
SHGetDataFromIDList() fails (which it will for library folders),
IPropertyStore::GetValue() will be used as a backup.

Note that to retrieve the IPropertyStore instance, the
IPropertyStoreFactory interface is retrieved first. That will fail in at
least the recycle bin folder, but that doesn't matter, as
SHGetDataFromIDList() will work there. The combination of both methods
should cover all folders.

Finally, as with SHGetDataFromIDList(), the property store will also
retrieve data cached in the pidl for the item. So there shouldn't be any
performance impact from using the property store when necessary.