Check return value when building attribute string
Generating the attribute string can fail for virtual items (e.g. the
virtual items contained in the desktop root folder).

Previously, the return value of the generation function wasn't being
checked. This meant that if the generation failed, the resulting string
would be uninitialized.

To fix this, the return value is now checked.