Don't show invalid dates or sizes in details view
Retrieving basic file information can (and will) fail for some items -
for example, the virtual items in the desktop root folder. When that
happens, the date/size columns should be blank, as there's nothing to
show.

Note that the sorting code has also been updated so that items without
dates/sizes will be sorted before other items. Additionally, the
grouping code has also been updated, to place items that have no data
for the relevant attribute in an "Unspecified" group.