Sort listview groups correctly
Previously, when grouping items, the groups would always be sorted by
name. While this is correct for most groups, it's not correct in some
cases.

For example, when grouping by ascending date (so that older items come
first) a group named "Yesterday" should be sorted before a group named
"Today". However, when sorting groups by name, the "Yesterday" group
would appear after the "Today" group.

To fix this, the following group types are now sorted based on the
relative positions of each group:

- Size
- Total size
- Date created
- Date modified
- Date accessed

Other group types will continue to be sorted by name.

This should mean that cases like the above are now handled correctly.

This closes #101.