Fix issue with sort mode verification
When navigating into a folder, the sort mode is verified (to ensure that
there's an associated column for that folder).

This check was previously performed by comparing the sort mode against
each column type. However, sort modes and column types are independent
and their values won't necessarily line up.

For example, ColumnType::FileSystem is 31, while SortMode::FileSystem is
34. Therefore, if you sorted by filesystem in the "This PC" folder, then
refreshed the folder, the sort mode would be reset.

To fix this, the verification function now compares the sort mode
against the sort mode for each column.