Add options to display mixed files and folders and sort names in non-natural sort order
This commit includes the necessary changes made in #53, though updated
to work with the latest version of the code.

Note that the sorting option only has an effect when sorting by name;
the other sorting options are unaffected. This matches the behaviour in
Windows Explorer, where once natural sorting is disabled, only the sort
by name functionality is affected. Other columns that contain numbers
will continue to be sorted in natural order.

This fixes #30 and fixes #46 and closes #53.