Update item group on modification
When a file or folder is modified, the group it's within might change.
For example, if items are being grouped by size, and the size of an item
changes, then the group that item is within will change.

Previously, this situation wasn't being handled at all; an item would
always remain within its original group. Now, the group an item is in
will be updated when that item is updated. Additionally, the header for
the group will be updated to indicate how many items are now in the
group.

This fixes #110.