Switch to new tab when using Ctrl + Shift + Double click
Previously, when double clicking a folder within the main pane, there
was no way to create a new tab and switch to it. Now, this can be done
by using Ctrl + Shift + Double click.

The following modifiers are available when double clicking:

- Alt + Double click: Shows item properties.
- Ctrl + Double click: Opens the folder in a background tab.
- Shift + Double click: Opens the folder in a new window.
- Ctrl + Shift + Double click: Opens the folder in a foreground tab.
- Double click: Opens the folder in the current tab.

Additionally, the same modifiers are used when selecting an item and
pressing enter.

This is part of the work for #303.