Add ability to choose whether tabs are opened in the background or foreground by default
An advanced option has been added that allows the user to choose whether
tabs are opened in the background or foreground by default.

Note that key combinations such as Shift + Middle click and Ctrl + Shift
+ Double click will result in the tab being opened in the opposite way
to the current preference.

For example, if tabs are set to open in the background and Shift +
Middle click is used, the new tab will be created in the foreground. On
the other hand, if tabs are set to open in the foreground and Shift +
Middle click is used, the new tab will be created in the background.

This closes #303.