Synchronise treeview for mapped drives
Previously, the treeview wasn't being synchronised for UNC folders or
folders contained on mapped drives. Now, only UNC folders are excluded
from synchronisation (due to the fact that they're contained within the
Network folder and that folder can take a significant amount of time to
enumerate).

Since the treeview enumerates folders on the main thread, any folders
that are slow to enumerate may cause hangs, including folders on mapped
drives.

This fixes #191.