Tweak thumbnails view
In thumbnails view, the LVS_EX_BORDERSELECT style is no longer set. This
makes the view more similar to that in Windows Explorer, when it
displays thumbnails.

Additionally, the icon spacing is no longer set to a custom value. The
default spacing works well and is less cramped than the custom spacing.