Set infotip string when there is no infotip to retrieve
It's possible for IQueryInfo::GetInfoTip() to return S_OK, even though
the pointer returned is null (because there's no infotip for the
specified item).

One situation in which I've seen that occur is in a Windows 7 VM, where
no infotip was returned for filesystem folders shown within the root
desktop folder.

In such a case, the infotip string should still be initialized (which it
is now).