Unfilter items after rename when necessary
Previously, items would be filtered out of the listview, if necessary,
when renaming. However, the reverse wasn't true. That is, if an item no
longer needed to be filtered after being renamed, it would remain hidden
regardless.

Now, items will be shown again after renaming if they no longer match
the filter that's set.