Restore selection when navigating to a history entry
When navigating to a history entry, the previous selection is now
restored. This happens when refreshing a folder, or navigating
back/forward.

Any currently selected items are stored within the current history entry
before a navigation occurs. Then, when a navigation to a history entry
has finished, the items contained within that entry will be selected
again.

This fixes #240 and fixes #212.