Cancel navigation when enumeration fails
Previously, when navigating to a directory that's not possible to
enumerate (e.g. because the user doesn't have the appropriate
permissions), Explorer++ would seemingly open the folder, though display
no items.

This behaviour has been updated so that Explorer++ will cancel the
navigation if the enumeration (or one of the steps before it) fails. In
such a case, the tab will remain in the original folder. This is also
what happens when a navigation fails in Windows Explorer.

One edge case is when a tab is created. In that situation, the tab will
need to be able to browse into a folder. Therefore, if the first
navigation of a new tab fails, the default directory will be loaded
instead.