Update a number of pieces of code to observe navigation committed signal
In various places, the navigation committed signal should be used
instead of the navigation started signal/navigation completed signal.

When the navigation is marked as committed, it's guaranteed that it will
proceed (something that's not necessarily true when the navigation
starts).

Additionally, it may not be necessary to wait until a navigation
completes to perform an update; when a navigation commits, a tab is
effectively now "in" that folder, even though none of the items have
loaded yet.