Retrieve IDataObject data from shell when copying files
Previously, the data stored within the IDataObject instance used when
copying to the clipboard was generated manually. Now, the IDataObject
instance is retrieved from the shell.

This has the advantage that not only does the data not have to be
generated manually, copying items from a virtual folder (e.g. the
virtual folder set up for a phone connected via USB) should now work,
since the appropriate formats are set up automatically. Additionally,
the IDataObject functionality doesn't have to be implemented.

Note that the reason the retrieved IDataObject instance isn't directly
used is that although it's possibly to retrieve the
IDataObjectAsyncCapability interface for the object and call
SetAsyncMode(), that call doesn't appear to actually enable asynchronous
copy, for whatever reason. Using asynchronous copy is important,
otherwise files will be copied on the main thread by the target.

Therefore, a wrapper object is used and it's that object that's copied
to the clipboard. The wrapper object simply delegates all IDataObject
interface calls, but enables asynchronous transfer.

This fixes #99 and fixes #80. This also fixes part of #144.