Retrieve IDataObject data from shell when dragging files
As with retrieving the shell IDataObject instance when copying files,
retrieving the instance when dragging and dropping means that the data
doesn't have to be generated manually.

This means that dragging and dropping from certain virtual folders (like
the virtual folder created for a phone connected via USB) should now
work.

Additionally, the code to start a drag has been moved into the
ShellBrowser class and simplified.

This fixes #144.