Support "New Folder" background context menu item
This menu item is shown on the background context menu for the virtual
folder used to display the contents of a phone connected via USB.

When the item is selected, the shell queries for the IShellView
interface. The SelectItem() method is then invoked to put the new folder
into rename mode.