Fix issue with locating renamed item
When an item is renamed, it's located using its pidl. Until recently,
the item was located by name, but this was changed to pidl in a recent
commit (e4bb1580918577dd967dc88011dd77e437152c03).

However, this was causing issues, since the parsing path of a directory
isn't necessarily unique. That means that building a simple pidl off of
a parsing path won't necessarily give you a pidl that matches any item
in the current directory.

For example, the desktop folder is located at
C:\Users\{Username}\Desktop, but it also appears at the root of the
namespace. These folders are technically independent items, even though
they both have the same parsing path - C:\Users\{Username}\Desktop.

Therefore, if you transform a path of the form
C:\Users\{Username}\Desktop\file into a simple pidl, the pidl you get
will only match the filesystem folder. It won't match any item in the
root desktop folder, even though the items in the root folder are
ultimately the same as the items in the filesystem folder.

Therefore, to fix this issue, the pidl that's created when an item is
renamed is now always relative to the parent folder.

The resolve link functionality has been updated in the same way.

This fixes #309.