Paste shortcuts by invoking "pastelink" verb
As with using the "paste" verb, this has the advantage that various
formats will be automatically supported. Additionally, if you paste the
same shortcut multiple times, a new shortcut will be created each time
(whereas before there would only ever be a single shortcut). This change
also means that the shortcut doesn't have to be manually created.