Update ShellBrowser to retrieve IDropTarget from shell when performing drop
This has the same advantages as the corresponding update to the
ShellTreeView class (made in 79e089fb917b38af6a7b465452171e913c9c5ba2).

As part of this update, it's now also possible to perform a drop on a
non-folder target (e.g. an executable or zip file).

Finally, the ability to rearrange files within a listview has
temporarily been removed.

This fixes #75 and fixes #297.